+++ tests/testthat/test-ggplot-lines.R
@@ -31,7 +31,7 @@ test_that("different colored lines become different colored traces", {
     geom_line()+
     scale_color_manual(values=c(y1="blue", y2="red"))
   info <- save_outputs(gg, "linetype-colors")
-  expect_equal(length(info$data), 2)
+  expect_equivalent(length(info$data), 2)
   expect_identical(info$data[[1]]$line$color, toRGB("blue"))
   n <- length(x)
   expect_identical(info$data[[1]]$y[1:n], y1)
@@ -69,9 +69,9 @@ test_that("Translates both dates and datetimes (with dynamic ticks) correctly", 
   
   # since these are dynamic ticks, let plotly.js generate the ticks
   axisType <- with(l$layout$xaxis, list(type, tickmode, autorange))
-  expect_equal(axisType, list("date", "auto", TRUE))
+  expect_equivalent(axisType, list("date", "auto", TRUE))
   axisType2 <- with(l2$layout$xaxis, list(type, tickmode, autorange))
-  expect_equal(axisType2, list("date", "auto", TRUE))
+  expect_equivalent(axisType2, list("date", "auto", TRUE))
   
   # range and data have been reverse transformed
   expect_is(l$layout$xaxis$range, "Date")
@@ -82,8 +82,8 @@ test_that("Translates both dates and datetimes (with dynamic ticks) correctly", 
   # check the hovertext
   dates1 <- sapply(strsplit(l$data[[1]]$text, br()), "[[", 1)
   dates2 <- sapply(strsplit(l2$data[[1]]$text, br()), "[[", 1)
-  expect_equal(paste("date:", d$date), dates1)
-  expect_equal(paste("date:", d2$date), dates2)
+  expect_equivalent(paste("date:", d$date), dates1)
+  expect_equivalent(paste("date:", d2$date), dates2)
 })
 
 test_that("geom_linerange() without a y aesthetic translates to a path", {
@@ -99,16 +99,16 @@ test_that("geom_linerange() without a y aesthetic translates to a path", {
   l <- plotly_build(p)$x
   
   expect_length(l$data, 1)
-  expect_equal(l$data[[1]]$type, "scatter")
-  expect_equal(
+  expect_equivalent(l$data[[1]]$type, "scatter")
+  expect_equivalent(
     l$data[[1]]$x,
     c(1, 1, NA, 2, 2, NA, 3, 3, NA, 4, 4, NA, 5, 5)
   )
-  expect_equal(
+  expect_equivalent(
     l$data[[1]]$y,
     c(0, 1, NA, 0, 2, NA, 0, 3, NA, 0, 4, NA, 0, 5)
   )
-  expect_equal(
+  expect_equivalent(
     unlist(l$data[[1]]$text),
     c(
       'x: 1<br />ymin: 0', 'x: 1<br />ymax: 1', NA, 'x: 2<br />ymin: 0',