+++ tests/testthat/test-ggplot-vline.R
@@ -11,8 +11,8 @@ test_that("second trace be the vline", {
   L <- save_outputs(p, "vline")
   l <- L$data[[2]]
   
-  expect_equal(length(L$data), 2)
-  expect_equal(l$x[1], 1.1)
+  expect_equivalent(length(L$data), 2)
+  expect_equivalent(l$x[1], 1.1)
   expect_true(l$y[1] <= 0)
   expect_true(l$y[2] >= 3.325)
   expect_true(l$mode == "lines")
@@ -23,7 +23,7 @@ test_that("vector xintercept results in multiple vertical lines", {
   p <- gg + geom_vline(xintercept = 1:2, colour = "blue", size = 3)
   
   L <- save_outputs(p, "vline-multiple")
-  expect_equal(length(L$data), 2)
+  expect_equivalent(length(L$data), 2)
   l <- L$data[[2]]
   xs <- unique(l$x)
   ys <- unique(l$y)